requirements.txt
asgiref==3.3.4
django-cors-headers
Django==3.2
django-filter==2.4.0
django-restframework==0.0.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
install==1.3.4
Pillow==8.2.0
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
boto3
django-storages

